docs: 폼 작성 가이드 내용 수정
